Chore: change amount type to Double

+++ book-service/src/main/java/com/ndrewcoding/bookservice/proxy/CambioProxy.java
@@ -5,12 +5,10 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-import java.math.BigDecimal;
-
 @FeignClient(name = "cambio-service", url = "localhost:8000")
 public interface CambioProxy {
 
     @GetMapping("/cambio-service/{amount}/{from}/{to}")
-    public Cambio getCambio(@PathVariable("amount") BigDecimal amount, @PathVariable("from") String from, @PathVariable("to") String to);
+    public Cambio getCambio(@PathVariable("amount") Double amount, @PathVariable("from") String from, @PathVariable("to") String to);
 
 }
